api/requirements.txt
alembic==1.6.5
boto3==1.17.88
botocore==1.20.88
click==8.0.1
Flask==2.0.1
Flask-Login==0.5.0
flask-marshmallow==0.14.0
Flask-Migrate==3.0.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==0.10.0
Mako==1.1.4
MarkupSafe==2.0.1
marshmallow==3.12.1
marshmallow-sqlalchemy==0.26.0
PyMySQL==1.0.2
python-dateutil==2.8.1
python-editor==1.0.4
s3transfer==0.4.2
six==1.16.0
SQLAlchemy==1.4.17
urllib3==1.26.5
Werkzeug==2.0.1
requests==2.25.1
uuid
PyJWT
datetime